+++ BUILD.gn
@@ -1231,7 +1231,6 @@ config("grpc_config") {
         "include/grpcpp/ext/health_check_service_server_builder_option.h",
         "include/grpcpp/generic/async_generic_service.h",
         "include/grpcpp/generic/generic_stub.h",
-        "include/grpcpp/generic/generic_stub_impl.h",
         "include/grpcpp/grpcpp.h",
         "include/grpcpp/health_check_service_interface.h",
         "include/grpcpp/impl/call.h",
