+++ gRPC-C++.podspec
@@ -89,7 +89,6 @@ Pod::Spec.new do |s|
                       'include/grpcpp/ext/health_check_service_server_builder_option.h',
                       'include/grpcpp/generic/async_generic_service.h',
                       'include/grpcpp/generic/generic_stub.h',
-                      'include/grpcpp/generic/generic_stub_impl.h',
                       'include/grpcpp/grpcpp.h',
                       'include/grpcpp/health_check_service_interface.h',
                       'include/grpcpp/impl/call.h',
